fix: create definitions for entities

+++ src/main/java/main/arbitrage/application/collector/dto/OrderbookDto.java
@@ -15,7 +15,7 @@ public class OrderbookDto {
     @Builder
     @Getter
     public static class OrderbookUnit {
-        private BigDecimal price;
-        private BigDecimal size;
+        private double price;
+        private double size;
     }
 }
